Revert "Experimenting with shipyard app"

+++ crates/mani/fabled_render/src/mesh/container/mesh_data.rs
@@ -1,9 +1,10 @@
 use crate::mesh::container::Vertex;
 use crate::mesh::Indices;
 
-#[derive(Debug, Clone)]
+#[derive(Debug)]
+#[repr(C, align(16))]
 pub struct Mesh {
     pub vertices: Vec<Vertex>,
     pub indices: Indices,
-    pub material_id: usize,
+    pub material_id: u32,
 }
